requirements.txt
fastapi>=0.115
uvicorn>=0.30
ollama>=0.4
httpx>=0.27
pyyaml>=6.0
python-dotenv>=1.0
